move blobsBundle validation logic into validateBlobsAndKzgCommitments
make validateBlobsAndKzgCommitments fork aware

add fulu to ForkBlobs

always call validateBlobsAndKzgCommitments on fulu: client should always verify
proofs
