+++ packages/beacon-node/src/chain/produceBlock/validateBlobsAndKzgCommitments.ts
@@ -1,16 +1,43 @@
 import {ExecutionPayload} from "@lodestar/types";
 import {BlobsBundle} from "../../execution/index.js";
+import { CELLS_PER_EXT_BLOB, ForkBlobs, ForkSeq } from "@lodestar/params";
+import { ckzg } from "../../util/kzg.js";
 
 /**
  * Optionally sanity-check that the KZG commitments match the versioned hashes in the transactions
  * https://github.com/ethereum/consensus-specs/blob/11a037fd9227e29ee809c9397b09f8cc3383a8c0/specs/eip4844/validator.md#blob-kzg-commitments
  */
 
-export function validateBlobsAndKzgCommitments(_payload: ExecutionPayload, blobsBundle: BlobsBundle): void {
-  // sanity-check that the KZG commitments match the blobs (as produced by the execution engine)
+export function validateBlobsAndKzgCommitments(fork: ForkBlobs, _payload: ExecutionPayload, blobsBundle: BlobsBundle): void {
   if (blobsBundle.blobs.length !== blobsBundle.commitments.length) {
     throw Error(
       `Blobs bundle blobs len ${blobsBundle.blobs.length} != commitments len ${blobsBundle.commitments.length}`
     );
+  }
+
+  if (ForkSeq[fork] >= ForkSeq.fulu) {
+    const expectedProofsLength = blobsBundle.blobs.length * CELLS_PER_EXT_BLOB;
+    if (blobsBundle.proofs.length !== expectedProofsLength) {
+      throw Error(
+        `Invalid proofs length for BlobsBundleV2 format: expected ${expectedProofsLength}, got ${blobsBundle.proofs.length}`
+      );
+    }
+
+    const commitmentBytes = blobsBundle.commitments.flatMap((commitment) => Array(CELLS_PER_EXT_BLOB).fill(commitment));
+    const cellIndices = Array.from({length: blobsBundle.blobs.length}).flatMap(() =>
+      Array.from({length: CELLS_PER_EXT_BLOB}, (_, i) => i)
+    );
+    const cells = blobsBundle.blobs.flatMap((blob) => ckzg.computeCells(blob));
+    const proofBytes = blobsBundle.proofs.flat();
+
+    try {
+      ckzg.verifyCellKzgProofBatch(commitmentBytes, cellIndices, cells, proofBytes);
+    } catch {
+      throw new Error("Error in verifyCellKzgProofBatch");
+    }
+  } else {
+    if (blobsBundle.proofs.length !== blobsBundle.blobs.length) {
+      throw new Error(`Invalid proofs length for BlobsBundleV1 format: expected ${blobsBundle.blobs.length}, got ${blobsBundle.proofs.length}`);
+    }
   }
 }
