+++ packages/beacon-node/test/unit/chain/produceBlock/validateBlobsAndKzgCommitments.test.ts
@@ -1,11 +1,12 @@
-import {describe, expect, it, beforeAll, afterEach} from "vitest";
+import {describe, expect, it, beforeAll} from "vitest";
 import {BlobsBundle} from "../../../../src/execution/index.js";
-import {validateBlobsBundle} from "../../../../src/chain/validation/blobsBundle.js";
+import {validateBlobsAndKzgCommitments} from "../../../../src/chain/produceBlock/validateBlobsAndKzgCommitments.js";
 import {ckzg, initCKZG, loadEthereumTrustedSetup} from "../../../../src/util/kzg.js";
 import {CELLS_PER_EXT_BLOB, ForkName} from "@lodestar/params";
 import {generateRandomBlob} from "../../../utils/kzg.js";
+import {ExecutionPayload} from "@lodestar/types";
 
-describe("validateBlobsBundle", () => {
+describe("validateBlobsAndKzgCommitments", () => {
   beforeAll(async () => {
     await initCKZG();
     loadEthereumTrustedSetup();
@@ -22,10 +23,17 @@ describe("validateBlobsBundle", () => {
       proofs,
     };
 
-    expect(() => validateBlobsBundle(ForkName.deneb, blobsBundle)).not.toThrow();
+    // Create a mock ExecutionPayload
+    const mockPayload = {
+      blockNumber: 1,
+      blockHash: new Uint8Array(32),
+      parentHash: new Uint8Array(32),
+    } as ExecutionPayload;
+
+    expect(() => validateBlobsAndKzgCommitments(ForkName.deneb, mockPayload, blobsBundle)).not.toThrow();
   });
 
-  it("should validate a valid V2 (fulu) blobs bundle", () => {
+  it("should validate a valid V2 (electra) blobs bundle", () => {
     const blobs = [generateRandomBlob(), generateRandomBlob()];
     const commitments = blobs.map((blob) => ckzg.blobToKzgCommitment(blob));
     const proofs = blobs.flatMap((blob) => ckzg.computeCellsAndKzgProofs(blob)[1]);
@@ -36,7 +44,14 @@ describe("validateBlobsBundle", () => {
       proofs,
     };
 
-    expect(() => validateBlobsBundle(ForkName.fulu, blobsBundle)).not.toThrow();
+    // Create a mock ExecutionPayload
+    const mockPayload = {
+      blockNumber: 1,
+      blockHash: new Uint8Array(32),
+      parentHash: new Uint8Array(32),
+    } as ExecutionPayload;
+
+    expect(() => validateBlobsAndKzgCommitments(ForkName.fulu, mockPayload, blobsBundle)).not.toThrow();
   });
 
   it("should throw if commitments and blobs lengths don't match", () => {
@@ -46,8 +61,15 @@ describe("validateBlobsBundle", () => {
       proofs: [],
     };
 
-    expect(() => validateBlobsBundle(ForkName.deneb, blobsBundle)).toThrow(
-      "Invalid BlobsBundle: commitments.length 1 != blobs.length 2"
+    // Create a mock ExecutionPayload
+    const mockPayload = {
+      blockNumber: 1,
+      blockHash: new Uint8Array(32),
+      parentHash: new Uint8Array(32),
+    } as ExecutionPayload;
+
+    expect(() => validateBlobsAndKzgCommitments(ForkName.deneb, mockPayload, blobsBundle)).toThrow(
+      "Blobs bundle blobs len 2 != commitments len 1"
     );
   });
 
@@ -60,7 +82,14 @@ describe("validateBlobsBundle", () => {
       proofs: [], // No proofs when we need one per blob
     };
 
-    expect(() => validateBlobsBundle(ForkName.deneb, blobsBundle)).toThrow(
+    // Create a mock ExecutionPayload
+    const mockPayload = {
+      blockNumber: 1,
+      blockHash: new Uint8Array(32),
+      parentHash: new Uint8Array(32),
+    } as ExecutionPayload;
+
+    expect(() => validateBlobsAndKzgCommitments(ForkName.deneb, mockPayload, blobsBundle)).toThrow(
       "Invalid proofs length for BlobsBundleV1 format: expected 1, got 0"
     );
   });
@@ -74,7 +103,14 @@ describe("validateBlobsBundle", () => {
       proofs: [new Uint8Array(48).fill(1)], // Only one proof when we need CELLS_PER_EXT_BLOB
     };
 
-    expect(() => validateBlobsBundle(ForkName.fulu, blobsBundle)).toThrow(
+    // Create a mock ExecutionPayload
+    const mockPayload = {
+      blockNumber: 1,
+      blockHash: new Uint8Array(32),
+      parentHash: new Uint8Array(32),
+    } as ExecutionPayload;
+
+    expect(() => validateBlobsAndKzgCommitments(ForkName.fulu, mockPayload, blobsBundle)).toThrow(
       `Invalid proofs length for BlobsBundleV2 format: expected ${CELLS_PER_EXT_BLOB}, got 1`
     );
   });
@@ -91,6 +127,15 @@ describe("validateBlobsBundle", () => {
       proofs,
     };
 
-    expect(() => validateBlobsBundle(ForkName.fulu, blobsBundle)).toThrow("Error in verifyCellKzgProofBatch");
+    // Create a mock ExecutionPayload
+    const mockPayload = {
+      blockNumber: 1,
+      blockHash: new Uint8Array(32),
+      parentHash: new Uint8Array(32),
+    } as ExecutionPayload;
+
+    expect(() => validateBlobsAndKzgCommitments(ForkName.fulu, mockPayload, blobsBundle)).toThrow(
+      "Error in verifyCellKzgProofBatch"
+    );
   });
 });
